# obsrvr-lake/stellar-query-api/go/docs/swagger.json
{
    "basePath": "/",
...
    "host": "gateway.withobsrvr.com",
    "info": {
        "contact": {
            "email": "support@obsrvr.com",
            "name": "OBSRVR Team",
            "url": "https://obsrvr.com"
        },
        "description": "API for querying Stellar blockchain data across Bronze, Silver, Semantic, and Gold data layers.\n\n## Data Layers\n- **Bronze**: Raw blockchain data (ledgers, transactions, operations, effects)\n- **Silver**: Enriched and processed data (accounts, assets, transfers, trades)\n- **Semantic**: Meaning-oriented views (activities, contracts, value flows)\n- **Gold**: Analytical views (snapshots, compliance reports)\n- **Index Plane**: Fast lookup services (transaction hash, contract events)\n\n## Amount Formatting\nAll amounts are returned in decimal format with 7 decimal places (Stellar stroops conversion).\nExample: Raw stroops `10000000` is formatted as `\"1.0000000\"`",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "Stellar Query API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/silver/address/{addr}/events": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "description": "Returns CAP-67 events where the address appears as sender (from) or receiver (to)",
                "parameters": [
                    {
                        "description": "Stellar account or contract address",
                        "in": "path",
                        "name": "addr",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Filter by event type: transfer, mint, burn",
                        "in": "query",
                        "name": "event_type",
                        "type": "string"
                    },
                    {
                        "description": "Filter by source: classic or soroban",
                        "in": "query",
                        "name": "source_type",
                        "type": "string"
                    },
                    {
                        "default": 20,
                        "description": "Max results (default: 20, max: 200)",
                        "in": "query",
                        "name": "limit",
                        "type": "integer"
                    },
                    {
                        "description": "Pagination cursor",
                        "in": "query",
                        "name": "cursor",
                        "type": "string"
                    },
                    {
                        "default": "desc",
                        "description": "Sort order",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "in": "query",
                        "name": "order",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Address events with pagination",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing address",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get events for an address",
                "tags": [
                    "Events"
                ]
            }
        },
...
        "/api/v1/silver/events": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "description": "Returns the unified CAP-67 event stream with optional filters for contract, event type, source type, and ledger range. Events are derived from token_transfers_raw with type inferred from from/to nullity.",
                "parameters": [
                    {
                        "description": "Filter by token contract ID (C...)",
                        "in": "query",
                        "name": "contract_id",
                        "type": "string"
                    },
                    {
                        "description": "Filter by event type: transfer, mint, burn",
                        "in": "query",
                        "name": "event_type",
                        "type": "string"
                    },
                    {
                        "description": "Filter by source: classic or soroban",
                        "in": "query",
                        "name": "source_type",
                        "type": "string"
                    },
                    {
                        "description": "Start of ledger range",
                        "in": "query",
                        "name": "start_ledger",
                        "type": "integer"
                    },
                    {
                        "description": "End of ledger range",
                        "in": "query",
                        "name": "end_ledger",
                        "type": "integer"
                    },
                    {
                        "default": 20,
                        "description": "Max results (default: 20, max: 200)",
                        "in": "query",
                        "name": "limit",
                        "type": "integer"
                    },
                    {
                        "description": "Pagination cursor from previous response",
                        "in": "query",
                        "name": "cursor",
                        "type": "string"
                    },
                    {
                        "default": "desc",
                        "description": "Sort order",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "in": "query",
                        "name": "order",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Unified events with pagination",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get unified CAP-67 event stream",
                "tags": [
                    "Events"
                ]
            }
        },
        "/api/v1/silver/events/by-contract": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "description": "Returns CAP-67 events filtered to a specific token contract",
                "parameters": [
                    {
                        "description": "Token contract ID (C...)",
                        "in": "query",
                        "name": "contract_id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Filter by event type: transfer, mint, burn",
                        "in": "query",
                        "name": "event_type",
                        "type": "string"
                    },
                    {
                        "description": "Filter by source: classic or soroban",
                        "in": "query",
                        "name": "source_type",
                        "type": "string"
                    },
                    {
                        "default": 20,
                        "description": "Max results (default: 20, max: 200)",
                        "in": "query",
                        "name": "limit",
                        "type": "integer"
                    },
                    {
                        "description": "Pagination cursor",
                        "in": "query",
                        "name": "cursor",
                        "type": "string"
                    },
                    {
                        "default": "desc",
                        "description": "Sort order",
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "in": "query",
                        "name": "order",
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Contract events with pagination",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing contract_id",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get events by contract",
                "tags": [
                    "Events"
                ]
            }
        },
...
        "/api/v1/silver/tx/{hash}/events": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "description": "Returns all CAP-67 events (transfers, mints, burns) for a specific transaction",
                "parameters": [
                    {
                        "description": "Transaction hash",
                        "in": "path",
                        "name": "hash",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Transaction events",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing transaction hash",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "additionalProperties": true,
                            "type": "object"
                        }
                    }
                },
                "summary": "Get events for a transaction",
                "tags": [
                    "Events"
                ]
            }
        },
...
    },
    "schemes": [
        "https"
    ],
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication. Format: \"Api-Key YOUR_API_KEY\"",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}